♻️ refactor: extract ZeroKMS functionality into dedicated module
Move ZeroKMS-related encryption/decryption logic from proxy/mod.rs into
a dedicated zerokms module for better code organization and maintainability.
Also fix clippy module_inception warnings.
